+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v1.1.0-1.16.5] - 2021-12-06
+### Added
+- Added new system for manually registering crafting table blocks that should be supported by Visual Workbench
+- Previous solution was too unreliable
+### Fixed
+- Fixed a client desync issue when FastWorkbench Minus Replacement mod is installed
+
 ## [v1.0.6-1.16.5] - 2021-11-28
 ### Fixed
 - Fixed an issue where shift-clicking multiple crafting results wouldn't work properly